requirements.txt
pandas
numpy
pydantic
pyyaml
streamlit
httpx
feedparser
pytest
